requirements.txt
opencv-python
plyer
requests
transformers
torch
torchvision
ultralytics
python-dotenv
deepface
tf-keras
tensorflow-cpu
mysql-connector-python
fastapi
uvicorn